+++ addon/query/js-adapter.js
@@ -2,6 +2,7 @@ import BaseAdapter from './base-adapter';
 import { SimplePredicate, ComplexPredicate, StringPredicate, DetailPredicate } from './predicate';
 import FilterOperator from './filter-operator';
 import Condition from './condition';
+import Information from '../utils/information';
 
 /**
  * Class of query language adapter that translates query object into JS function which
@@ -150,26 +151,27 @@ export function buildProjection(query) {
 /**
  * Builds function for filtering array of objects using predicate.
  *
- * @param predicate Predicate for array of objects.
+ * @param predicate Predicate for filtering array of objects.
  * @returns {Function}
  */
-export function buildFilter(predicate) {
+function buildFilter(predicate) {
   let b1 = predicate instanceof SimplePredicate;
   let b2 = predicate instanceof StringPredicate;
   let b3 = predicate instanceof DetailPredicate;
 
   if (b1 || b2 || b3) {
-    return getSimpleFilterFunction(predicate);
+    let filterFunction = getAttributeFilterFunction(predicate);
+    return getFilterFunctionAnd([filterFunction]);
   }
 
   if (predicate instanceof ComplexPredicate) {
     let filterFunctions = predicate.predicates.map(getAttributeFilterFunction);
     switch (predicate.condition) {
       case Condition.And:
-        return getComplexFilterFunctionAnd(filterFunctions);
+        return getFilterFunctionAnd(filterFunctions);
 
       case Condition.Or:
-        return getComplexFilterFunctionOr(filterFunctions);
+        return getFilterFunctionOr(filterFunctions);
 
       default:
         throw new Error(`Unsupported condition '${predicate.condition}'.`);
@@ -180,52 +182,31 @@ export function buildFilter(predicate) {
 }
 
 /**
- * Returns function for filtering array of objects using simple or string predicate.
- * Result function filters array of objects and returns those, for which
- * an attribute filter function returned `true`.
+ * Returns function for checkign single object using predicate.
  *
- * @param {Query.SimplePredicate|Query.ComplexPredicate} predicate Predicate for array of objects.
- * @returns {Function}
- */
-function getSimpleFilterFunction(predicate) {
-  let attributeFilter = getAttributeFilterFunction(predicate);
-  return function (data) {
-    let result = [];
-    data.forEach(i => {
-      if (attributeFilter(i)) {
-        result.push(i);
-      }
-    });
-    return result;
-  };
-}
-
-/**
- * Returns function for filtering single attribute of an object using predicate.
- *
- * @param {Query.SimplePredicate|Query.ComplexPredicate} predicate Predicate for an attribute.
- * @returns {Function} Function for filtering single attribute.
+ * @param {Query.BasePredicate} predicate Predicate for an attribute.
+ * @returns {Function} Function for checkign single object.
  */
 export function getAttributeFilterFunction(predicate) {
   if (predicate instanceof SimplePredicate) {
     switch (predicate.operator) {
       case FilterOperator.Eq:
-        return function (i) { return i[predicate.attributeName] === predicate.value; };
+        return (i) => getValue(i, predicate.attributePath) === predicate.value;
 
       case FilterOperator.Neq:
-        return function (i) { return i[predicate.attributeName] !== predicate.value; };
+        return (i) => getValue(i, predicate.attributePath) !== predicate.value;
 
       case FilterOperator.Le:
-        return function (i) { return i[predicate.attributeName] < predicate.value; };
+        return (i) => getValue(i, predicate.attributePath) < predicate.value;
 
       case FilterOperator.Leq:
-        return function (i) { return i[predicate.attributeName] <= predicate.value; };
+        return (i) => getValue(i, predicate.attributePath) <= predicate.value;
 
       case FilterOperator.Ge:
-        return function (i) { return i[predicate.attributeName] > predicate.value; };
+        return (i) => getValue(i, predicate.attributePath) > predicate.value;
 
       case FilterOperator.Geq:
-        return function (i) { return i[predicate.attributeName] >= predicate.value; };
+        return (i) => getValue(i, predicate.attributePath) >= predicate.value;
 
       default:
         throw new Error(`Unsupported filter operator '${predicate.operator}'.`);
@@ -233,27 +214,29 @@ export function getAttributeFilterFunction(predicate) {
   }
 
   if (predicate instanceof StringPredicate) {
-    return function (i) { return i[predicate.attributeName].indexOf(predicate.containsValue) > -1; };
+    return (i) => getValue(i, predicate.attributePath).indexOf(predicate.containsValue) > -1;
   }
 
   if (predicate instanceof DetailPredicate) {
     let detailFilter = buildFilter(predicate.predicate);
     if (predicate.isAll) {
       return function (i) {
-        if (!i[predicate.detailName]) {
+        let detail = getValue(i, predicate.detailPath);
+        if (!detail) {
           return false;
         }
 
-        let result = detailFilter(i[predicate.detailName]);
-        return result.length === i[predicate.detailName].length;
+        let result = detailFilter(detail);
+        return result.length === detail.length;
       };
     } else if (predicate.isAny) {
       return function (i) {
-        if (!i[predicate.detailName]) {
+        let detail = getValue(i, predicate.detailPath);
+        if (!detail) {
           return false;
         }
 
-        let result = detailFilter(i[predicate.detailName]);
+        let result = detailFilter(detail);
         return result.length > 0;
       };
     } else {
@@ -261,7 +244,61 @@ export function getAttributeFilterFunction(predicate) {
     }
   }
 
+  if (predicate instanceof ComplexPredicate) {
+    let filterFunctions = predicate.predicates.map(getAttributeFilterFunction);
+    switch (predicate.condition) {
+      case Condition.And:
+        return function (i) {
+          let check = true;
+          for (let funcIndex = 0; funcIndex < filterFunctions.length; funcIndex++) {
+            check &= filterFunctions[funcIndex](i);
+            if (!check) {
+              break;
+            }
+          }
+
+          return check;
+        };
+
+      case Condition.Or:
+        return function (i) {
+          let check = false;
+          for (let funcIndex = 0; funcIndex < filterFunctions.length; funcIndex++) {
+            check |= filterFunctions[funcIndex](i);
+            if (check) {
+              break;
+            }
+          }
+
+          return check;
+        };
+
+      default:
+        throw new Error(`Unsupported condition '${predicate.condition}'.`);
+    }
+  }
+
   throw new Error(`Unsupported predicate '${predicate}'.`);
+}
+
+/**
+ Loads value from object by specified attribute path.
+
+ @param {Object} item Object for load value.
+ @param {String} attributePath The path to the attribute.
+ @returns {*|undefined} Value of attribute or `undefined`.
+ */
+function getValue(item, attributePath) {
+  let attributes = Information.parseAttributePath(attributePath);
+  let search = item;
+  for (let i = 0; i < attributes.length; i++) {
+    search = search[attributes[i]];
+    if (!search) {
+      return undefined;
+    }
+  }
+
+  return search;
 }
 
 /**
@@ -273,7 +310,7 @@ export function getAttributeFilterFunction(predicate) {
  * @param {Function[]} filterFunctions Array of attribute filter functions.
  * @returns {Function} Complex filter function for `or` condition.
  */
-function getComplexFilterFunctionAnd(filterFunctions) {
+function getFilterFunctionAnd(filterFunctions) {
   return function (data) {
     let result = [];
     for (let itemIndex = 0; itemIndex < data.length; itemIndex++) {
@@ -303,7 +340,7 @@ function getComplexFilterFunctionAnd(filterFunctions) {
  * @param {Function[]} filterFunctions Array of attribute filter functions.
  * @returns {Function} Complex filter function for `or` condition.
  */
-function getComplexFilterFunctionOr(filterFunctions) {
+function getFilterFunctionOr(filterFunctions) {
   return function (data) {
     let result = [];
     for (let itemIndex = 0; itemIndex < data.length; itemIndex++) {