+++ tests/unit/query/indexeddb-adapter-test.js
@@ -12,16 +12,16 @@ import startApp from '../../helpers/start-app';
 const app = startApp();
 const store = app.__container__.lookup('service:store');
 const databasePrefix = 'testDB2';
-const schema = 'Id,Name,Surname,Age';
+const schema = 'Id,Name,Surname,Age,Manager.Name';
 const modelName = 'employee';
 
 module('query');
 
 test('adapter | indexeddb | without predicate', (assert) => {
   let data = [
-    { Id: 1, Name: 'A', Surname: 'X', Age: 10 },
-    { Id: 2, Name: 'A', Surname: 'Y', Age: 11 },
-    { Id: 3, Name: 'B', Surname: 'Z', Age: 15 }
+    { Id: 1 },
+    { Id: 2 },
+    { Id: 3 }
   ];
 
   let builder = new QueryBuilder(store, modelName);
@@ -37,9 +37,9 @@ test('adapter | indexeddb | without predicate', (assert) => {
 
 test('adapter | indexeddb | simple predicate | eq', (assert) => {
   let data = [
-    { Id: 1, Name: 'A', Surname: 'X', Age: 10 },
-    { Id: 2, Name: 'B', Surname: 'Y', Age: 11 },
-    { Id: 3, Name: 'B', Surname: 'Z', Age: 12 }
+    { Id: 1, Name: 'A' },
+    { Id: 2, Name: 'B' },
+    { Id: 3, Name: 'B' }
   ];
 
   let builder = new QueryBuilder(store, modelName).where('Name', FilterOperator.Eq, 'B');
@@ -49,6 +49,22 @@ test('adapter | indexeddb | simple predicate | eq', (assert) => {
     assert.equal(result.length, 2);
     assert.equal(result[0].Id, 2);
     assert.equal(result[1].Id, 3);
+  });
+});
+
+test('adapter | js | simple predicate | eq | master field', function (assert) {
+  let data = [
+    { Id: 1, Manager: { Name: 'X' } },
+    { Id: 2 },
+    { Id: 3, Manager: { Name: 'Y' } }
+  ];
+
+  let builder = new QueryBuilder(store, modelName).where('Manager.Name', FilterOperator.Eq, 'Y');
+
+  executeTest(data, builder.build(), assert, (result) => {
+    assert.ok(result);
+    assert.equal(result.length, 1);
+    assert.equal(result[0].Id, 3);
   });
 });
 
@@ -139,9 +155,9 @@ test('adapter | indexeddb | simple predicate | geq', (assert) => {
 
 test('adapter | indexeddb | string predicate | contains', (assert) => {
   let data = [
-    { Id: 1, Name: 'A', Surname: 'X', Age: 10, Country: 'Argentina' },
-    { Id: 2, Name: 'B', Surname: 'Y', Age: 11, Country: 'Paragwaj' },
-    { Id: 3, Name: 'C', Surname: 'Z', Age: 12, Country: 'Russia' }
+    { Id: 1, Country: 'Argentina' },
+    { Id: 2, Country: 'Paragwaj' },
+    { Id: 3, Country: 'Russia' }
   ];
 
   let sp1 = new StringPredicate('Country').contains('i');
@@ -155,7 +171,25 @@ test('adapter | indexeddb | string predicate | contains', (assert) => {
   });
 });
 
-test('adapter | indexeddb | detail predicate | all | with simple predicate', (assert) => {
+test('adapter | indexeddb | string predicate | contains | master field', (assert) => {
+  let data = [
+    { Id: 1, Country: { Name: 'Argentina' } },
+    { Id: 2, Country: { Name: 'Paragwaj' } },
+    { Id: 3, Country: { Name: 'Russia' } }
+  ];
+
+  let sp1 = new StringPredicate('Country.Name').contains('i');
+  let builder = new QueryBuilder(store, modelName).where(sp1);
+
+  executeTest(data, builder.build(), assert, (result) => {
+    assert.ok(result);
+    assert.equal(result.length, 2);
+    assert.equal(result[0].Id, 1);
+    assert.equal(result[1].Id, 3);
+  });
+});
+
+test('adapter | indexeddb | detail predicate | all | simple predicate', (assert) => {
   let data = [
     { Id: 1, Tags: [{ Name: 'Tag1' }] },
     { Id: 2 },
@@ -172,7 +206,24 @@ test('adapter | indexeddb | detail predicate | all | with simple predicate', (as
   });
 });
 
-test('adapter | indexeddb | detail predicate | any | with simple predicate', (assert) => {
+test('adapter | indexeddb | detail predicate | all | simple predicate | master field', (assert) => {
+  let data = [
+    { Id: 1, Tags: [{ Creator: { Name: 'X' } }] },
+    { Id: 2 },
+    { Id: 3 }
+  ];
+
+  let dp = new DetailPredicate('Tags').all(new SimplePredicate('Creator.Name', FilterOperator.Eq, 'X'));
+  let builder = new QueryBuilder(store, modelName).where(dp);
+
+  executeTest(data, builder.build(), assert, (result) => {
+    assert.ok(result);
+    assert.equal(result.length, 1);
+    assert.equal(result[0].Id, 1);
+  });
+});
+
+test('adapter | indexeddb | detail predicate | any | simple predicate', (assert) => {
   let data = [
     { Id: 1, Tags: [{ Name: 'Tag1' }, { Name: 'Tag3' }] },
     { Id: 2, Tags: [{ Name: 'Tag3' }, { Name: 'Tag2' }] },
@@ -190,7 +241,25 @@ test('adapter | indexeddb | detail predicate | any | with simple predicate', (as
   });
 });
 
-test('adapter | indexeddb | detail predicate | all | with complex predicate', (assert) => {
+test('adapter | indexeddb | detail predicate | any | simple predicate | master field', (assert) => {
+  let data = [
+    { Id: 1, Tags: [{ Creator: { Name: 'X' } }, { Creator: { Name: 'Z' } }] },
+    { Id: 2, Tags: [{ Creator: { Name: 'Z' } }, { Creator: { Name: 'Y' } }] },
+    { Id: 3, Tags: [{ Creator: { Name: 'Y' } }, { Creator: { Name: 'X' } }] }
+  ];
+
+  let dp = new DetailPredicate('Tags').any(new SimplePredicate('Creator.Name', FilterOperator.Eq, 'X'));
+  let builder = new QueryBuilder(store, modelName).where(dp);
+
+  executeTest(data, builder.build(), assert, (result) => {
+    assert.ok(result);
+    assert.equal(result.length, 2);
+    assert.equal(result[0].Id, 1);
+    assert.equal(result[1].Id, 3);
+  });
+});
+
+test('adapter | indexeddb | detail predicate | all | complex predicate', (assert) => {
   let data = [
     { Id: 1, Tags: [{ Name: 'Tag1' }, { Name: 'Tag3' }] },
     { Id: 2, Tags: [{ Name: 'Tag3' }, { Name: 'Tag2' }] },
@@ -210,7 +279,27 @@ test('adapter | indexeddb | detail predicate | all | with complex predicate', (a
   });
 });
 
-test('adapter | indexeddb | detail predicate | any | with complex predicate', (assert) => {
+test('adapter | indexeddb | detail predicate | all | complex predicate | master field', (assert) => {
+  let data = [
+    { Id: 1, Tags: [{ Creator: { Name: 'X' } }, { Creator: { Name: 'Z' } }] },
+    { Id: 2, Tags: [{ Creator: { Name: 'Z' } }, { Creator: { Name: 'Y' } }] },
+    { Id: 3, Tags: [{ Creator: { Name: 'Y' } }, { Creator: { Name: 'X' } }] }
+  ];
+
+  let sp1 = new SimplePredicate('Creator.Name', FilterOperator.Eq, 'X');
+  let sp2 = new SimplePredicate('Creator.Name', FilterOperator.Eq, 'Z');
+  let cp1 = new ComplexPredicate(Condition.Or, sp1, sp2);
+  let dp = new DetailPredicate('Tags').all(cp1);
+  let builder = new QueryBuilder(store, modelName).where(dp);
+
+  executeTest(data, builder.build(), assert, (result) => {
+    assert.ok(result);
+    assert.equal(result.length, 1);
+    assert.equal(result[0].Id, 1);
+  });
+});
+
+test('adapter | indexeddb | detail predicate | any | complex predicate', (assert) => {
   let data = [
     { Id: 1, Tags: [{ Name: 'Tag4' }, { Name: 'Tag3' }] },
     { Id: 2, Tags: [{ Name: 'Tag3' }, { Name: 'Tag1' }] },
@@ -219,6 +308,27 @@ test('adapter | indexeddb | detail predicate | any | with complex predicate', (a
 
   let sp1 = new SimplePredicate('Name', FilterOperator.Eq, 'Tag1');
   let sp2 = new SimplePredicate('Name', FilterOperator.Eq, 'Tag2');
+  let cp1 = new ComplexPredicate(Condition.Or, sp1, sp2);
+  let dp = new DetailPredicate('Tags').any(cp1);
+  let builder = new QueryBuilder(store, modelName).where(dp);
+
+  executeTest(data, builder.build(), assert, (result) => {
+    assert.ok(result);
+    assert.equal(result.length, 2);
+    assert.equal(result[0].Id, 2);
+    assert.equal(result[1].Id, 3);
+  });
+});
+
+test('adapter | indexeddb | detail predicate | any | complex predicate', (assert) => {
+  let data = [
+    { Id: 1, Tags: [{ Creator: { Name: 'M' } }, { Creator: { Name: 'Z' } }] },
+    { Id: 2, Tags: [{ Creator: { Name: 'Z' } }, { Creator: { Name: 'X' } }] },
+    { Id: 3, Tags: [{ Creator: { Name: 'Y' } }, { Creator: { Name: 'A' } }] }
+  ];
+
+  let sp1 = new SimplePredicate('Creator.Name', FilterOperator.Eq, 'X');
+  let sp2 = new SimplePredicate('Creator.Name', FilterOperator.Eq, 'Y');
   let cp1 = new ComplexPredicate(Condition.Or, sp1, sp2);
   let dp = new DetailPredicate('Tags').any(cp1);
   let builder = new QueryBuilder(store, modelName).where(dp);
@@ -270,6 +380,31 @@ test('adapter | indexeddb | complex predicate | or', (assert) => {
     assert.equal(result.length, 2);
     assert.equal(result[0].Surname, 'X');
     assert.equal(result[1].Surname, 'Z');
+  });
+});
+
+test('adapter | indexeddb | complex predicate | with nested complex predicate', function (assert) {
+  // Arrange.
+  let data = [
+    { Id: 1,  Name: 'A', Surname: 'X', Age: 10 },
+    { Id: 2,  Name: 'B', Surname: 'Y', Age: 11 },
+    { Id: 3,  Name: 'C', Surname: 'Z', Age: 12 }
+  ];
+
+  let sp1 = new SimplePredicate('Name', FilterOperator.Eq, 'A');
+  let sp2 = new SimplePredicate('Surname', FilterOperator.Eq, 'Z');
+  let cp1 = new ComplexPredicate(Condition.Or, sp1, sp2);
+
+  let sp3 = new SimplePredicate('Age', FilterOperator.Eq, 12);
+  let cp2 = new ComplexPredicate(Condition.And, cp1, sp3);
+
+  // Act && Assert.
+  let builder = new QueryBuilder(store, modelName).where(cp2);
+
+  executeTest(data, builder.build(), assert, (result) => {
+    assert.ok(result);
+    assert.equal(result.length, 1);
+    assert.equal(result[0].Surname, 'Z');
   });
 });