+++ tests/unit/query/odata-adapter-test.js
@@ -23,6 +23,15 @@ test('adapter | odata | id', function (assert) {
   assert.equal(url, baseUrl + '/Customers(42)');
 });
 
+test('adapter | odata | simple predicate | master field', function (assert) {
+  // Arrange && Act.
+  let builder = new QueryBuilder(store, 'customer').where('manager.First Name', FilterOperator.Eq, 'Vasya');
+  let url = adapter.getODataFullUrl(builder.build());
+
+  // Assert.
+  assert.equal(url, `${baseUrl}/Customers?$filter=manager/first-name eq 'Vasya'`);
+});
+
 test('adapter | odata | simple predicate | eq', function (assert) {
   // Arrange && Act.
   let builder = new QueryBuilder(store, 'customer').where('firstName', FilterOperator.Eq, 'Vasya');
@@ -190,6 +199,23 @@ test('adapter | odata | complex predicate', function (assert) {
 
   // Assert.
   assert.equal(url, `${baseUrl}/Customers?$filter=first-name eq 'Vasya' or last-name eq 'Ivanov' or age eq 10`);
+});
+
+test('adapter | odata | complex predicate | with nested complex predicate', function (assert) {
+  // Arrange.
+  let sp1 = new SimplePredicate('firstName', FilterOperator.Eq, 'Vasya');
+  let sp2 = new SimplePredicate('lastName', FilterOperator.Eq, 'Ivanov');
+  let cp1 = new ComplexPredicate(Condition.Or, sp1, sp2);
+
+  let sp3 = new SimplePredicate('age', FilterOperator.Eq, 10);
+  let cp2 = new ComplexPredicate(Condition.And, cp1, sp3);
+
+  // Act.
+  let builder = new QueryBuilder(store, 'customer').where(cp2);
+  let url = adapter.getODataFullUrl(builder.build());
+
+  // Assert.
+  assert.equal(url, `${baseUrl}/Customers?$filter=(first-name eq 'Vasya' or last-name eq 'Ivanov') and age eq 10`);
 });
 
 test('adapter | odata | order', function (assert) {