+++ CHANGELOG.md
@@ -3,3 +3,5 @@
 ## [Unreleased]
 ### Added
 - Predicates for details with any / all functionality.
+### Fixed
+- Using nested complex predicates for all adapters.